requirements.txt
django-model-utils==2.2